Make `size` return an `int` instead of a `double`
Just saw it in [Tyler's fork](https://github.com/calcmogul/allwpilib/commit/c277dd26c714839e8b267169364e8916009c1272#diff-147b1b2f2d1b675c5a40b533567eecea28e981ee76339cccc7ffbb1e25d30d3fL34)

+++ wpiutil/src/main/java/edu/wpi/first/util/CircularBuffer.java
@@ -31,7 +31,7 @@ public class CircularBuffer {
    *
    * @return number of elements in buffer
    */
-  public double size() {
+  public int size() {
     return m_length;
   }
